+++ CHANGELOG.md
@@ -495,6 +495,7 @@
 
 - Serve CSV downloads encoded in UTF-8, prefixed with a byte order mark
 - Show textarea content with wrapper HTML on the Activity details page
+- BEIS users can create a programme-level activity associated with a source fund
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-31...HEAD
 [release-31]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-30...release-31